+++ src/client/src/pages/detail/form/borehole/boreholePanelInterfaces.ts
@@ -1,6 +1,6 @@
 import { BoreholeV2 } from "../../../../api/borehole.ts";
 import { BoreholeFile } from "../../../../api/file/fileInterfaces.ts";
-import { WorkflowV2 } from "../../../../api/workflow.ts";
+import { WorkflowV2 } from "../workflow/workflow.ts";
 
 export interface BoreholeProps {
   borehole: BoreholeV2;