+++ src/client/src/utils.ts
@@ -1,4 +1,4 @@
-export function capitalizeFirstLetter(text: string) {
+export function capitalizeFirstLetter(text: string | undefined) {
   if (!text) return "";
   return text.charAt(0).toUpperCase() + text.slice(1);
 }