+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added the photos tab to the borehole attachments.
 - Added photo viewer to the borehole sidepanel.
 - Added data collection to detail page.
+- Automatic conversion from meters above sea level (MASL) to measured depth (MD) in hydro module.
 
 ### Changed
 - Upgraded to React 19.